Add `govuk_sidekiq/testing`
This is a replacement for `require "sidekiq/testing"`. It will require
the sidekiq testing library and set up the testing middleware.

The way we use Sidekiq middleware is that
`APIHeaders::ClientMiddleware` adds an extra argument to the job, and
`APIHeaders::ServerMiddleware` removes it.

But Sidekiq by default does not load server middleware[1], which causes
the extra argument to be added, but never removed. Sidekiq then calls
the job with the extra argument, resulting in lots of `ArgumentError:
wrong number of arguments (given 2, expected 1).`

[1]
https://github.com/mperham/sidekiq/wiki/Testing#testing-server-middlewar
e