Use GovukStatsd for statsd logging
It is expected that all the apps that should be using this gem should also be
using govuk_app_config - thus this is added as a dependency of this gem. A
minimum of version 1.1 is required as this is the first version
compatible with the statsd methods required by sidekiq-statsd.

We then use GovukStatsd as the client for statsd which will allow us to
make use of the consistent namespacing that provides across our apps.