Remove REDIS_HOST/REDIS_PORT configuration appraoch
This settles on REDIS_URL as the means to configure this gem, this was
chosen as it is one natively understood by the Redis gem [1] and thus
more common.

This choice has been removed as part of a GOV.UK effort to remove the
confusion that can be caused by having two different configuration
approaches used across the platform.

It also updates the SidekiqInitializer#setup_sidekiq method to not
require a redis config since this is not necessarily needed when
REDIS_URL is set (the Redis gem will pick this up automatically).

[1]: https://github.com/redis/redis-rb