Redefine sidekiq requirement to not install beta version
Currently, bundler will install version `7.0.0.beta1` of sidekiq as it
considers this to be less than `7`. We do not want this to happen.

In https://github.com/alphagov/govuk_sidekiq/pull/84, we changed the
requirement from `">= 6.5.12", "< 7"` to `"~> 6.5"`. This has a side
effect of some of our applications ending up on an earlier version of
sidekiq than `6.5.12` (there was a security fix in `6.5.10`).

Therefore in https://github.com/alphagov/govuk_sidekiq/pull/85 we
changed the requirement to be `"> 6.5.10", "< 7"`, however this once
again resulted in bundler installing `7.0.0.beta`.

Updating the requirements again to be even more explicit, so we always
install a version newer than `6.5.12` but never a version that isn't
`6.5.x`.