Fix `reconnect_attempts` not working with Redis 4.8
Our apps are currently using Redis 4.8 due to a constraint from the
version of Sidekiq we are using.

However in a previous commit, we used Redis 5 syntax to specify a new
value for `reconnect_attempts`. This is not compatible with Redis 4.8.

Therefore updating the syntax to match that expected for Redis 4.8 [1]
and explicitly stating the version of Redis we expect (since Redis 5
doesn't support the legacy syntax).

1: https://github.com/redis/redis-rb/tree/v4.8.1?tab=readme-ov-file#reconnections