Don't verify TLS
This is needed when running a Heroku Redis instance other than mini
where you are only given a rediss:// URL and self signed certificate.

This turns off verification of the certificate.

Sidekiq fails to run without this [1].

This is a quick proof of concept so I can link to this branch for a
heroku example, as I couldn't work out how to disable the initializer in
the Railstie so I could run the below code in my app directly.

[1]: https://github.com/sidekiq/sidekiq/wiki/Heroku#redis