Reduce the time window for Sidekiq retries
The previous code meant that if there was a Sidekiq connection error
there would be a `sleep(15)` called before any attempt to reconnect to
Redis. For a client application serving a web request and adding to the
Sidekiq this 15 seconds is a very long time to wait.

This is substantially longer than the gem default values which are a
0.25s wait twice [1]

The motivation for making this change was finding that in a Heroku
environment (which we are currently using for demoing an app) it is
frequent that we receive an intermittment 15 second wait to serve a
request. The Redis connection must be less reliable on the cheaper
hardware.

[1]: https://github.com/redis-rb/redis-client/blob/8dfe1f65208482bb30150dd0009254ac77fe4776/lib/redis_client/sentinel_config.rb#L7-L8