Add a couple of longer Redis reconnection retries
Until recently our retries were set to [15, 30, 45, 60] (seconds). The
change to the current snappier settings has exposed how surprisingly
long it can take for new Redis instances to be available after they've
been killed. And that's resulted in a lot of noise in Sentry when Redis
instances are restarted across multiple apps and those apps can't
reconnect within 5 seconds.

We don't think that this is a great long-term fix for that, but it
should reduce the noise in Sentry while we look into our Redis setup.